requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
cryptography==3.4.6
Django==3.2
django-js-asset==1.2.2
django-mptt==0.11.0
factory-boy==3.2.0
Faker==8.1.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
isort==5.7.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.1.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
